% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capin_metadata.R
\name{capin_metadata}
\alias{capin_metadata}
\title{file metadata}
\usage{
capin_metadata(x, metadata)
}
\arguments{
\item{x}{(character) a file path}

\item{metadata}{metadata something or other}
}
\value{
file path for the metadata file, containing
RDF
}
\description{
associate metadata with a file
}
\examples{
library(contentid)
f <- system.file("extdata", "vostok.icecore.co2", package = "contentid")
x <- capin(f)

# create some metadata file
library(rdflib)
r <- rdf()
rdf_add(r,
  subject = "http://www.dajobe.org/",
  predicate = "http://purl.org/dc/elements/1.1/language",
  object = "en")
rdf_add(r,
  subject = "b0",
  predicate = "http://purl.org/dc/elements/1.1/title",
  object = "Dave Beckett's Home Page")
rdf_add(r,
  subject = "b0",
  predicate = "http://purl.org/dc/elements/1.1/title",
  object = "Dave Beckett's Home Page")
rdf_add(r,
  subject = "b0",
  predicate = "http://purl.org/dc/elements/1.1/creator",
  object = "Dave Beckett")
# write metadata to a file
mfile <- capin_metadata(x, metadata=r)
# read back the metadata file
rdflib::rdf_parse(mfile)
}
