% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file}
\alias{file}
\title{file}
\usage{
file(x)
}
\arguments{
\item{x}{(character) a file path}
}
\description{
file
}
\examples{
# a file that's already on the web
x <- system.file("extdata", "vostok.icecore.co2", package = "contentid")
file(x)
url <- 'https://cdiac.ess-dive.lbl.gov/ftp/trends/co2/vostok.icecore.co2'
file(url)

# a file not on the web
w <- tempfile(fileext=".csv")
df <- data.frame(a = letters, b = LETTERS)
write.csv(df, file = w)
res <- file(w)
res
contentid::query_sources(res)
}
