% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{file_metadata}
\alias{file_metadata}
\title{file metadata}
\usage{
file_metadata(x, metadata)
}
\arguments{
\item{x}{(character) a file path}

\item{metadata}{metadata something or other}
}
\value{
file path for the metadata file, containing
RDF
}
\description{
associate metadata with a file
}
\examples{
library(contentid)
f <- system.file("extdata", "vostok.icecore.co2", package = "contentid")
x <- file(f)

# create some metadata file
library(rdflib)
r <- rdf()
z <- rdf_add(r,
  subject = "http://www.dajobe.org/",
  predicate = "http://purl.org/dc/elements/1.1/language",
  object = "en")
file_metadata(x, metadata=z)
}
