% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_files.R
\name{same_files}
\alias{same_files}
\title{check if two files are the same}
\usage{
same_files(x, y)
}
\arguments{
\item{x, y}{two file paths}
}
\value{
boolean, whether files have the same content or not,
based on comparing output of \code{\link[contentid:content_id]{contentid::content_id()}} with \code{\link[=identical]{identical()}}
}
\description{
use case: user wants to compare if contents of two files are the same
}
\examples{
x <- tempfile()
writeLines("foo bar\n", x)
y <- tempfile()
file.copy(x, y)
same_files(x, y)
}
