% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capin.R
\name{capin}
\alias{capin}
\title{file}
\usage{
capin(x, meta = NULL)
}
\arguments{
\item{x}{(character) a file path}

\item{meta}{either a file path (character), rdf objet from package
\pkg{rdflib}, or \code{NULL}}
}
\description{
file
}
\examples{
# a file that's already on the web
x <- system.file("extdata", "vostok.icecore.co2", package = "contentid")
capin(x)
url <- 'https://cdiac.ess-dive.lbl.gov/ftp/trends/co2/vostok.icecore.co2'
capin(url)

# a file not on the web
w <- tempfile(fileext=".csv")
df <- data.frame(a = letters, b = LETTERS)
write.csv(df, file = w)
res <- capin(w)
res
contentid::query_sources(res)

# associate metadata with the data
x <- system.file("extdata", "vostok.icecore.co2", package = "contentid")
m <- system.file("examples/rdf_eg1.rdf", package = "capins")
capin(x, meta = m)
}
